backend/requirements.txt
Flask==2.2.2
Flask-CORS==3.0.10
googletrans==4.0.0rc1
Werkzeug==2.2.2
python-dotenv